% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/termInfoExtractorFunction.R
\name{getAncestors}
\alias{getAncestors}
\title{Get list of ancestor terms}
\usage{
getAncestors(obj = NULL, term = NULL, ordering = FALSE)
}
\arguments{
\item{obj}{List obj (created by reader function)}

\item{term}{GO id character}

\item{ordering}{boolean to choose if the ancestors are ordered by level (from high level to low level)}
}
\value{
A vector object including the ancestors of term.
}
\description{
Get list of ancestor terms
}
\examples{
\dontrun{

data(go)


getAncestors(go, "GO:0007005",ordering=TRUE)
}
}
