% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{getChildren}
\alias{getChildren}
\title{Get list of child terms}
\usage{
getChildren(obj = NULL, term = NULL)
}
\arguments{
\item{obj}{List obj (created by reader function)}

\item{term}{GO id character}

\item{ordering}{boolean to choose if the children are ordered by level (from high level to low level)}
}
\value{
A vector object including the children of term.
}
\description{
Get list of child terms
}
\examples{
\dontrun{

data(go)


getChildren(go, "GO:0007005",ordering=TRUE)
}
}
