% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{getDescendantsWithLimit}
\alias{getDescendantsWithLimit}
\title{Get list of descendants terms but with a limited distance}
\usage{
getDescendantsWithLimit(obj = NULL, term = NULL, ordering = FALSE, limit)
}
\arguments{
\item{obj}{ontology R6 class}

\item{term}{GO id character}

\item{ordering}{boolean to choose if the descendants are ordered by level (from high level to low level)}

\item{limit}{number of maximal steps for exploring the go structure.}
}
\value{
A vector object including the descendants of term.
}
\description{
Get list of descendants terms but with a limited distance
}
\examples{
\dontrun{

data(go)


getDescendantsWithLimit(go, "GO:0007005",ordering=TRUE,limit=2)
}
}
