% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/termInfoExtractorFunction.R
\name{getParent}
\alias{getParent}
\title{Get list of parent terms}
\usage{
getParent(obj = NULL, term = NULL)
}
\arguments{
\item{obj}{List obj (created by reader function)}

\item{term}{GO id character}

\item{ordering}{boolean to choose if the parents are ordered by level (from high level to low level)}
}
\value{
A vector object including the parents of term.
}
\description{
Get list of parent terms
}
\examples{
\dontrun{

data(go)


getParent(go, "GO:0007005",ordering=TRUE)
}
}
