% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{go}
\alias{go}
\title{GO index}
\format{
List of lists and vectors
}
\usage{
data(go)
}
\description{
\code{ontologyReader} object encapsulating structure of the Gene Ontology (GO) comprising a \code{list} of lists/vectors of properties of GO terms indexed by term ID
GO version (Downloaded 04/2020)
}
