% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontologyReader.R
\name{GS2}
\alias{GS2}
\title{Compute GS2 (time-consuming)}
\usage{
GS2(geneset, annotaton, go)
}
\arguments{
\item{geneset}{vector string including genes}

\item{go}{gene Ontology object}

\item{annotation}{List including a vector of GO terms for each gene}
}
\value{
numerical value.
}
\description{
Compute GS2 (time-consuming)
}
