% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6class.R
\name{ontology}
\alias{ontology}
\title{ontology R6 class}
\description{
This object include the main caracteristics to get into a biological ontology
such as Gene Ontontology. With the object, you should be able to get the Gene
Ontology term information: name, depth, information content, and taxonomy relations.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{ids}}{Character vector including the ids of a given ontology.}

\item{\code{alternativeIDs}}{List of possible alternative ids for the ontology ids.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ontology$new()}}
\item \href{#method-print}{\code{ontology$print()}}
\item \href{#method-children}{\code{ontology$children()}}
\item \href{#method-parents}{\code{ontology$parents()}}
\item \href{#method-ancestors}{\code{ontology$ancestors()}}
\item \href{#method-descendants}{\code{ontology$descendants()}}
\item \href{#method-depth}{\code{ontology$depth()}}
\item \href{#method-ic}{\code{ontology$ic()}}
\item \href{#method-all_ic}{\code{ontology$all_ic()}}
\item \href{#method-top}{\code{ontology$top()}}
\item \href{#method-name}{\code{ontology$name()}}
\item \href{#method-clone}{\code{ontology$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
To create a new instance of this [ontology][R6::R6Class] class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ontology$new(ontologyList)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ontologyList}}{A list of three elements:\cr
  \itemize{
  \item{}{ontologyList$name include a vector of ids of an ontology (e.g. GO:0008150
  is a id in the Gene Ontology).}
  \item{}{ontologyList$termOBJ include a list of term information (name, depth, children,
  parents, ancestors, descendants, ...).}
  \item{}{ontologyList$alternativeIds includes a list of terms and their alternative id.}

  }
  This list is generated by the [ontologyReader::ontologyReader()] method.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
Print this [ontology][R6::R6Class] class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ontology$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-children"></a>}}
\if{latex}{\out{\hypertarget{method-children}{}}}
\subsection{Method \code{children()}}{
To extract the children of a given term.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ontology$children(term)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{term}}{term id character.\cr}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Ontology id children vector of the give term.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-parents"></a>}}
\if{latex}{\out{\hypertarget{method-parents}{}}}
\subsection{Method \code{parents()}}{
To extract the parents of a given term.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ontology$parents(term)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{term}}{term id character.\cr}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Ontology id parents vector of the give term.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ancestors"></a>}}
\if{latex}{\out{\hypertarget{method-ancestors}{}}}
\subsection{Method \code{ancestors()}}{
To extract the ancestors of a given term.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ontology$ancestors(term)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{term}}{term id character.\cr}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Ontology id ancestors vector of the give term.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-descendants"></a>}}
\if{latex}{\out{\hypertarget{method-descendants}{}}}
\subsection{Method \code{descendants()}}{
To extract the descendants of a given term.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ontology$descendants(term)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{term}}{term id character.\cr}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Ontology id descendants vector of the give term.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-depth"></a>}}
\if{latex}{\out{\hypertarget{method-depth}{}}}
\subsection{Method \code{depth()}}{
To extract depth position (longest path from root) of a given term.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ontology$depth(term)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{term}}{term id character.\cr}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
term depth position.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ic"></a>}}
\if{latex}{\out{\hypertarget{method-ic}{}}}
\subsection{Method \code{ic()}}{
To extract the information content (IC) presented by Mazandu and Mulder (2013) for a given GO term.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ontology$ic(term)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{term}}{term id character.\cr}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
IC value of the given term.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-all_ic"></a>}}
\if{latex}{\out{\hypertarget{method-all_ic}{}}}
\subsection{Method \code{all_ic()}}{
To extract the information content (IC) presented by Mazandu and Mulder (2013) for all GO terms.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ontology$all_ic()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
IC value of the given term.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-top"></a>}}
\if{latex}{\out{\hypertarget{method-top}{}}}
\subsection{Method \code{top()}}{
To extract the top term associated to the given term.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ontology$top(term)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{term}}{term id character.\cr}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
top term.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-name"></a>}}
\if{latex}{\out{\hypertarget{method-name}{}}}
\subsection{Method \code{name()}}{
To extract term name.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ontology$name(term)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{term}}{term id character.\cr}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
term name.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ontology$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
