% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{ontologyReader}
\alias{ontologyReader}
\title{Read ontology task}
\usage{
ontologyReader(file)
}
\arguments{
\item{file}{obo file}
}
\value{
A ontoClass object.
}
\description{
Read ontology task
}
\examples{
\dontrun{

onto <- ontologyReader("go.obo")


}

}
