% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{getDescendants}
\alias{getDescendants}
\title{Get list of descendants terms}
\usage{
getDescendants(obj = NULL, term = NULL, ordering = FALSE)
}
\arguments{
\item{obj}{List obj (created by reader function)}

\item{term}{GO id character}

\item{ordering}{boolean to choose if the descendants are ordered by level (from high level to low level)}
}
\value{
A vector object including the descendants of term.
}
\description{
Get list of descendants terms
}
\examples{
\dontrun{

data(go)


getDescendats(go, "GO:0007005",ordering=TRUE)
}
}
